requirements.txt
nltk==3.7
langchain==0.1.1
huggingface-hub==0.11.0
pinecone-client==2.2.4
python-dotenv==1.0.0
streamlit==1.29.0